pyproject.toml
[project]
name = "modernblog"
version = "1.0.0"
description = "A self-hosted, modern blogging platform with a beautiful UI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    {name = "Patryk Klatka", email = "contact@klatka.it"}
]
keywords = ["blog", "cms", "markdown", "self-hosted"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content :: Content Management System",
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-slugify>=8.0.0",
    "pyjwt>=2.8.0",
    "pillow>=10.4.0",
    "click>=8.1.0",
    "rich>=13.0.0",
]

[project.urls]
Homepage = "https://github.com/pklatka/modernblog"
Repository = "https://github.com/pklatka/modernblog"
Documentation = "https://github.com/pklatka/modernblog#readme"

[project.scripts]
modernblog = "modernblog:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["modernblog"]
exclude = ["modernblog/frontend"]

[tool.hatch.build.targets.wheel.force-include]
"modernblog/frontend/dist" = "modernblog/static/dist"

[tool.hatch.build.targets.wheel.hooks.custom]
path = "scripts/hatch_build.py"


[dependency-groups]
dev = ["ruff>=0.8.0"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["modernblog"]

